Make ui test that are run-pass and do not test the compiler itself library tests
