--- src/test/ui/assert-eq-trailing-comma.rs
@@ -1,5 +1,0 @@
-// run-pass
-
-fn main() {
-    assert_eq!(1, 1,);
-}
