--- src/test/ui/assert-escape.rs
@@ -1,5 +1,0 @@
-// run-pass
-
-fn main() {
-    assert!(r#"☃\backslash"#.contains("\\"));
-}
