--- src/test/ui/assert-ne-trailing-comma.rs
@@ -1,5 +1,0 @@
-// run-pass
-
-fn main() {
-    assert_ne!(1, 2,);
-}
